+++ src/CoordinationIdFactory.php
@@ -4,15 +4,12 @@ declare(strict_types = 1);
 
 namespace byrokrat\id;
 
-/**
- * Create coordination id objects from raw id string
- */
 class CoordinationIdFactory implements IdFactoryInterface
 {
     use Helper\IdFactoryDecoratorTrait;
 
-    protected function createNewInstance(string $raw): IdInterface
+    protected function createNewInstance(string $raw, \DateTimeInterface $atDate = null): IdInterface
     {
-        return new CoordinationId($raw);
+        return new CoordinationId($raw, $atDate);
     }
 }