+++ src/FailingIdFactory.php
@@ -4,13 +4,12 @@ declare(strict_types = 1);
 
 namespace byrokrat\id;
 
-/**
- * Id factory that always fails
- */
+use byrokrat\id\Exception\UnableToCreateIdException;
+
 class FailingIdFactory implements IdFactoryInterface
 {
-    public function createId(string $raw): IdInterface
+    public function createId(string $raw, \DateTimeInterface $atDate = null): IdInterface
     {
-        throw new Exception\UnableToCreateIdException("Unable to create ID for number '{$raw}'");
+        throw new UnableToCreateIdException("Unable to create ID for number '{$raw}'");
     }
 }