+++ src/FakeIdFactory.php
@@ -4,15 +4,12 @@ declare(strict_types = 1);
 
 namespace byrokrat\id;
 
-/**
- * Create fake id objects from raw id string
- */
 class FakeIdFactory implements IdFactoryInterface
 {
     use Helper\IdFactoryDecoratorTrait;
 
-    protected function createNewInstance(string $raw): IdInterface
+    protected function createNewInstance(string $raw, \DateTimeInterface $atDate = null): IdInterface
     {
-        return new FakeId($raw);
+        return new FakeId($raw, $atDate);
     }
 }