+++ src/IdFactoryInterface.php
@@ -2,15 +2,16 @@
 
 namespace byrokrat\id;
 
-/**
- * The base id factory interface
- */
+use byrokrat\id\Exception\UnableToCreateIdException;
+
 interface IdFactoryInterface
 {
     /**
      * Create ID object from raw id string
      *
-     * @throws Exception\UnableToCreateIdException If unable to create id
+     * @param string $raw The raw id to parse
+     * @param \DateTimeInterface $atDate Optional date when parsing takes place, defaults to today
+     * @throws UnableToCreateIdException If unable to create id
      */
-    public function createId(string $raw): IdInterface;
+    public function createId(string $raw, \DateTimeInterface $atDate = null): IdInterface;
 }