+++ src/NullIdFactory.php
@@ -4,12 +4,9 @@ declare(strict_types = 1);
 
 namespace byrokrat\id;
 
-/**
- * Create null id objects
- */
 class NullIdFactory implements IdFactoryInterface
 {
-    public function createId(string $raw): IdInterface
+    public function createId(string $raw, \DateTimeInterface $atDate = null): IdInterface
     {
         return new NullId;
     }