+++ src/OrganizationIdFactory.php
@@ -4,14 +4,11 @@ declare(strict_types = 1);
 
 namespace byrokrat\id;
 
-/**
- * Create organization id objects from raw id string
- */
 class OrganizationIdFactory implements IdFactoryInterface
 {
     use Helper\IdFactoryDecoratorTrait;
 
-    protected function createNewInstance(string $raw): IdInterface
+    protected function createNewInstance(string $raw, \DateTimeInterface $atDate = null): IdInterface
     {
         return new OrganizationId($raw);
     }