+++ tests/FakeIdFactoryTest.php
@@ -19,7 +19,7 @@ class FakeIdFactoryTest extends \PHPUnit\Framework\TestCase
         $id = $this->createMock(IdInterface::CLASS);
 
         $decorated = $this->prophesize(IdFactoryInterface::CLASS);
-        $decorated->createId('foo')->willReturn($id);
+        $decorated->createId('foo', null)->willReturn($id);
 
         $this->assertSame(
             $id,